+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Other:
 
 - Add keyword arguments to `Unicode::DisplayWidth.of`. If you are using a hash
   with overwrite value, be sure to put it in curly braces.
+- Using third parameter or explicit hash as fourth parameter is deprecated,
+  please migrate to the keyword arguments API
 - Gem raises ArgumentError for ambiguous values other than 1 or 2
 - Performance optimizations